Tweaks to multi protocol pacing
