+++ myokit/_protocol.py
@@ -1,5 +1,5 @@
 #
-# Defines the python classes that represent a pacing protocol.
+# Defines the python classes that represent pacing protocols.
 #
 # This file is part of Myokit.
 # See http://myokit.org for copyright, sharing, and licensing details.
@@ -10,6 +10,8 @@ from __future__ import print_function, unicode_literals
 import myokit
 
 import numpy as np
+
+from bisect import bisect_right
 
 
 class Protocol(object):
@@ -850,3 +852,93 @@ class NotAnUnbrokenSequenceError(myokit.MyokitError):
     """ Error raised exclusively by is_unbroken_sequence_exception(). """
     pass
 
+
+class TimeSeriesProtocol(object):
+    """
+    Represents a pacing protocol as a sequence of time value pairs and an
+    interpolation method (currently only linear interpolation is supported).
+
+    A 1D time-series should be given as input. During the simulation, the value
+    of the pacing variable will be determined by interpolating between the two
+    nearest points in the series. If the simulation time is outside the bounds
+    of the time-series, the first or last value in the series will be used.
+
+    Protocols can be compared with ``==``, which will check if the sequence of
+    time value pairs is the same, and the interpolation method is the same.
+    Protocols can be serialised with ``pickle``.
+
+    """
+
+    def __init__(self, times, values, method=None):
+        super(TimeSeriesProtocol, self).__init__()
+
+        if len(times) != len(values):
+            raise ValueError('Times and values array must have same size.')
+        times_tpl, values_tpl = zip(*[
+            (float(t), float(v)) for t, v in sorted(zip(times, values))
+        ])
+        self._times = list(times_tpl)
+        self._values = list(values_tpl)
+
+        if method is None:
+            self._method = 'linear'
+        else:
+            self._method = str(method).lower()
+            if self._method not in ('linear', ):
+                raise ValueError('Unknown interpolation method: ' + self._method)
+
+    def __eq__(self, other):
+        if self is other:
+            return True
+        if not isinstance(other, TimeSeriesProtocol):
+            return False
+        return (
+            self._method == other._method
+            and self._values == other._values
+            and self._times == other._times
+        )
+
+    def __getstate__(self):
+        return {
+            'times': self._times,
+            'values': self._values,
+            'method': self._method,
+        }
+
+    def __setstate__(self, values):
+        self._times = values['times']
+        self._values = values['values']
+        self._method = values['method']
+
+    def clone(self):
+        """
+        Returns a clone of this protocol.
+        """
+        return TimeSeriesProtocol(self._times, self._values, self._method)
+
+    def pace(self, t):
+        """
+        Returns the value of the pacing variable at time ``t``.
+        """
+        if t < self._times[0]:
+            return self._values[0]
+        if t > self._times[-1]:
+            return self._values[-1]
+        i = bisect_right(self._times, t) - 1
+        if i == len(self._times) - 1:
+            return self._values[i]
+        return self._values[i] + (t - self._times[i]) * (
+            self._values[i + 1] - self._values[i]
+        ) / (self._times[i + 1] - self._times[i])
+
+    def times(self):
+        """
+        Returns a list of the times in this protocol.
+        """
+        return self._times
+
+    def values(self):
+        """
+        Returns a list of the values in this protocol.
+        """
+        return self._values
